renames failed with special characters
I discovered this bug when I accidentally renamed a file with a \ in the filename, and when I tried to rename the file again to remove the \, I got a "Invalid value for cast/constructor" error.  This fixes the problem by encoding the rename parameter before passing it to the rename function.