Fix handling of XQueryServlet errors
A subtle change between eXist 5.2.0 and 5.3.0 meant that the fix in #261 caused every eXide query to raise an error for users of eXist 5.2.0 but not for those of 5.3.0. #261 introduced a check of request:get-data() to detect errors thrown by the XQueryServlet that lacked descriptions and had silently failed in earlier versions of eXide. In eXist 5.2.0, the call to request:get-data() here returns an empty string when there is no XQueryServlet error. Under eXist 5.3.0-SNAPSHOT, the call to request:get-data here returns an empty sequence. This commit makes eXide’s error reporting introduced in #261 impervious to this difference between eXist 5.2.0 and 5.3.0 and adds some comments to aid in maintenance.